aws/dynamoDB/cdk/requirements.txt
aws-cdk-lib==2.43.1
constructs>=10.0.0,<11.0.0

